requirements.txt
websockets>=11.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pyyaml>=6.0
